% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Print Hello World}
\usage{
hello()
}
\arguments{
\item{data}{the data that is used in this function}
}
\value{
this is what will be returned
}
\description{
test description
}
\examples{
#this is example code of the function
hello()
}
