% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{init_params}
\alias{init_params}
\title{Initialize CausalModels Package}
\usage{
init_params(outcome, treatment, covariates, data, simple = F)
}
\arguments{
\item{outcome}{the outcome variable of interest (must be continuous).}

\item{treatment}{the treatment with the causal effect of interest on the outcome.}

\item{covariates}{a list/vector of covariate names to be use for confounding adjustment.}

\item{data}{a data frame containing the variables in the model.}

\item{simple}{a boolean indicator to build default formula with interactions.
If true, interactions will be excluded. If false, interactions will be included. By
default, simple is set to false.}
}
\description{
This function is required to be run first before any other function can run.
This will set within the package the global outcome, treatment, and covariate functions for each model to use.
}
