% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_regression.R
\name{outcome_regression}
\alias{outcome_regression}
\title{Propensity Scores}
\usage{
outcome_regression(
  data,
  f = NA,
  contrasts = list(),
  simple = pkg.env$simple,
  ...
)
}
\arguments{
\item{data}{a data frame containing the variables in the model.
This should be the same data used in \code{\link[init_params]{init_params}}.}

\item{f}{(optional) an object of class "formula" that overrides the default parameter}

\item{simple}{a boolean indicator to build default formula with interactions.
If true, interactions will be excluded. If false, interactions will be included. By
default, simple is set to false.}

\item{...}{additional arguments that may be passed to the underlying \code{\link[stats:glm]{glm}} model.}

\item{family}{the family to be used in the general linear model.
By default, this is set to \code{\link[stats:binomial]{binomial}}
NOTE: if this is changed, the outcome of the model may not be the probabilities and the results will not be valid.}
}
\value{
\code{outcome_regression} returns an object of \code{\link[base::class]{class} "outcome_regression"}

The function \code{summary} can be used to obtain and print a summary of the underlying glm model.

An object of class \code{"outcome_regression"} is a list containing the following:

\tabular{ll}{
 \code{call} \tab the matched call. \cr
 \tab \cr
 \code{formula} \tab the formula used in the model. \cr
 \tab \cr
 \code{model} \tab the underlying glm model. \cr
 \tab \cr
 \code{p.scores} \tab the estimated propensity scores.\cr
}
}
\description{
`outcome_regression` builds a logistic regression with the target as the treatment variable
and the covariates as the independent variables.
}
\examples{
library(causaldata)
library(multcomp)

data(nhefs)
nhefs.nmv <- nhefs[which(!is.na(nhefs$wt82)),]
nhefs.nmv$qsmk <- as.factor(nhefs.nmv$qsmk)

confounders <- c("sex", "race", "age", "education", "smokeintensity",
                     "smokeyrs", "exercise", "active", "wt71")

init_params(wt82_71, qsmk,
            covariates = confounders,
            data = nhefs.nmv)

p.score <- outcome_regression(nhefs.nmv)
p.score
}
