% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_matching.R
\name{propensity_matching}
\alias{propensity_matching}
\title{Propensity Scores}
\usage{
propensity_matching(
  data,
  f = NA,
  simple = pkg.env$simple,
  p.scores = NA,
  p.simple = pkg.env$simple,
  type = "strata",
  grp.width = 0.1,
  quant = T,
  ...
)
}
\arguments{
\item{data}{a data frame containing the variables in the model.
This should be the same data used in \code{\link[init_params]{init_params}}.}

\item{f}{(optional) an object of class "formula" that overrides the default parameter}

\item{simple}{a boolean indicator to build default formula with interactions.
If true, interactions will be excluded. If false, interactions will be included. By
default, simple is set to false.}

\item{...}{additional arguments that may be passed to the underlying \code{\link[stats:glm]{glm}} model.}

\item{family}{the family to be used in the general linear model.
By default, this is set to \code{\link[stats:binomial]{binomial}}
NOTE: if this is changed, the outcome of the model may not be the probabilities and the results will not be valid.}
}
\value{
\code{propensity_matching} returns an object of \code{\link[base::class]{class} "propensity_matching"}

The function \code{summary} can be used to obtain and print a summary of the underlying glm model.

An object of class \code{"propensity_matching"} is a list containing the following:

\tabular{ll}{
 \code{call} \tab the matched call. \cr
 \tab \cr
 \code{formula} \tab the formula used in the model. \cr
 \tab \cr
 \code{model} \tab the underlying glm model. \cr
 \tab \cr
 \code{p.scores} \tab the estimated propensity scores.\cr
}
}
\description{
`propensity_matching` builds a logistic regression with the target as the treatment variable
and the covariates as the independent variables.
}
\examples{
library(causaldata)
library(multcomp)

data(nhefs)
nhefs.nmv <- nhefs[which(!is.na(nhefs$wt82)),]
nhefs.nmv$qsmk <- as.factor(nhefs.nmv$qsmk)

confounders <- c("sex", "race", "age", "education", "smokeintensity",
                     "smokeyrs", "exercise", "active", "wt71")

init_params(wt82_71, qsmk,
            covariates = confounders,
            data = nhefs.nmv)

p.score <- propensity_matching(nhefs.nmv)
p.score
}
